+++ src/main/java/com/bsodsoftware/merbackend/rest/TMRest.java
@@ -1,18 +1,21 @@
 package com.bsodsoftware.merbackend.rest;
 
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.bsodsoftware.merbackend.services.ObjetivoAprendizajeService;
+import com.bsodsoftware.merbackend.services.SecurityService;
 import com.bsodsoftware.merbackend.services.TareaMatematicaService;
 import com.bsodsoftware.merbackend.services.to.OaTmDto;
 import com.bsodsoftware.merbackend.services.to.ResponseDTO;
@@ -28,15 +31,24 @@ public class TMRest {
 	@Autowired
 	private ObjetivoAprendizajeService oaService;
 	
+	@Autowired
+	private SecurityService securityService;
+	
 	@PostMapping("/save")
 	@CrossOrigin
 	@ResponseBody
-	public ResponseDTO guardarTm(@RequestBody TMDTO tmDto) {
+	public ResponseDTO guardarTm(@RequestBody TMDTO tmDto,@RequestHeader("Authorization") String token) {
 		ResponseDTO ret = new ResponseDTO();
 		
 		try {
-			tmService.guardarTm(tmDto, 1L);
-			ret.setCodigo(200);
+			Long idUsuario = securityService.validateToken(token);
+			if (!idUsuario.equals(-1L)) {
+				tmService.guardarTm(tmDto, idUsuario);
+				ret.setCodigo(200);
+			} else {
+				ret.setCodigo(500);
+				ret.setComentario("Usuario de token no encontrado");
+			}
 		} catch (Exception ex) {
 			ret.setCodigo(500);
 			ret.setComentario(ex.getMessage());
@@ -48,20 +60,34 @@ public class TMRest {
 	@GetMapping("getAllTms")
 	@CrossOrigin
 	@ResponseBody
-	public List<TMDTO> getTms() {
-		return tmService.getTms();
+	public List<TMDTO> getTms(@RequestHeader("Authorization") String token) {
+		List<TMDTO> ret = null;
+		try {
+			Long idUsuario = securityService.validateToken(token);
+			if (!idUsuario.equals(-1L)) {
+				ret = tmService.getTms();
+			}
+		} catch (Exception ex) {
+			ex.printStackTrace();
+		}
+		return ret;
 	}
 	
 	@GetMapping("deleteTm")
 	@CrossOrigin
 	@ResponseBody
-	public ResponseDTO deleteTm(@RequestParam Long id) {
+	public ResponseDTO deleteTm(@RequestParam Long id,@RequestHeader("Authorization") String token) {
 		ResponseDTO ret = new ResponseDTO();
-		
 		try {
-			tmService.delete(id);
-			ret.setCodigo(200);
-			ret.setComentario("Eliminado correctamente");
+			Long idUsuario = securityService.validateToken(token);
+			if (!idUsuario.equals(-1L)) {
+				tmService.delete(id);
+				ret.setCodigo(200);
+				ret.setComentario("Eliminado correctamente");
+			} else {
+				ret.setCodigo(500);
+				ret.setComentario("Usuario de token no encontrado");
+			}
 		} catch (Exception ex) {
 			ret.setCodigo(500);
 			ret.setComentario(ex.getLocalizedMessage());
@@ -73,14 +99,32 @@ public class TMRest {
 	@GetMapping("getAllOaTms")
 	@CrossOrigin
 	@ResponseBody
-	public List<OaTmDto> getOaTms() {
-		return oaService.getOasTms();
+	public List<OaTmDto> getOaTms(@RequestHeader("Authorization") String token) {
+		List<OaTmDto> ret = null;
+		try {
+			Long idUsuario = securityService.validateToken(token);
+			if (!idUsuario.equals(-1L)) {
+				ret = oaService.getOasTms();
+			}
+		}catch (Exception ex) {
+			ex.printStackTrace();
+		}
+		return ret;
 	}
 	
 	@GetMapping("getTm")
 	@CrossOrigin
 	@ResponseBody
-	public TMDTO getTm(Long id) {
-		return tmService.getTmDto(id);
+	public TMDTO getTm(Long id,@RequestHeader("Authorization") String token) {
+		TMDTO ret = new TMDTO();
+		try {
+			Long idUsuario = securityService.validateToken(token);
+			if (!idUsuario.equals(-1L)) {
+				ret = tmService.getTmDto(id);
+			}
+		}catch (Exception ex) {
+			ex.printStackTrace();
+		}
+		return ret;
 	}
 }
